Fixed a silly mistake: changes to config.inc.php reverted
What I originally wanted to include in the config.inc.php were the
settings for two new tables that weren't there in the previous version.
So I blindly copied the config from my running container into the
repo and pushed it. That was a stupid thing to do as the vars in the
config were replaced by the actual values that I was using in my
container!